Maintenance
Broke out object and syntax extraction scripts into objects_syntax,
better commenting for jazzkey
